Fix button styling in top menu bar and cleanup/restructure code a bit
