+++ src/panels/Menu/TopMenuBar/TopMenuBar.tsx
@@ -1,20 +1,10 @@
 import { Dispatch, SetStateAction } from 'react';
-import { Button, Group, Kbd, Menu, Text } from '@mantine/core';
-import { modals } from '@mantine/modals';
+import { alpha, Group } from '@mantine/core';
 
-import { useGetBoolPropertyValue, useOpenSpaceApi } from '@/api/hooks';
-import {
-  AddFileIcon,
-  ConsoleIcon,
-  ExitAppIcon,
-  SaveIcon,
-  VisibilityIcon
-} from '@/icons/icons';
-
-import { HelpMenu } from './HelpMenu';
-import { MenuWrapper } from './MenuWrapper';
-import { NewWindowMenu } from './NewWindowMenu';
-import { TaskBarMenuChoices } from './TaskBarMenuChoices';
+import { FileMenu } from './Menus/FileMenu';
+import { HelpMenu } from './Menus/HelpMenu';
+import { ViewMenu } from './Menus/ViewMenu';
+import { WindowsMenu } from './Menus/WindowsMenu';
 
 interface TopMenuBarProps {
   visibleMenuItems: string[];
@@ -22,80 +12,14 @@ interface TopMenuBarProps {
 }
 
 export function TopMenuBar({ visibleMenuItems, setVisibleMenuItems }: TopMenuBarProps) {
-  const luaApi = useOpenSpaceApi();
-
-  const [isConsoleVisible, setIsConsoleVisible] =
-    useGetBoolPropertyValue('LuaConsole.IsVisible');
-
-  function toggleLuaConsole() {
-    setIsConsoleVisible(!isConsoleVisible);
-  }
-
-  function toggleShutdown() {
-    return modals.openConfirmModal({
-      title: 'Confirm action',
-      children: <Text>Are you sure you want to quit OpenSpace?</Text>,
-      labels: { confirm: 'Quit', cancel: 'Cancel' },
-      confirmProps: { color: 'red' },
-      onConfirm: () => luaApi?.toggleShutdown()
-    });
-  }
-
   return (
-    <Group h={30} w={'100%'} bg={'var(--mantine-color-gray-filled)'}>
-      <MenuWrapper>
-        <Menu.Target>
-          <Button size={'xs'} color={'gray'}>
-            File
-          </Button>
-        </Menu.Target>
-        <Menu.Dropdown>
-          <Menu.Item leftSection={<AddFileIcon />}>Add Asset</Menu.Item>
-          <Menu.Item
-            onClick={toggleLuaConsole}
-            leftSection={<ConsoleIcon />}
-            rightSection={<Kbd>~</Kbd>}
-          >
-            Toggle Console
-          </Menu.Item>
-          <Menu.Divider />
-          <Menu.Item
-            onClick={toggleShutdown}
-            leftSection={<ExitAppIcon />}
-            rightSection={<Kbd>Esc</Kbd>}
-          >
-            Quit OpenSpace
-          </Menu.Item>
-        </Menu.Dropdown>
-      </MenuWrapper>
-
-      <MenuWrapper closeOnItemClick={false}>
-        <Menu.Target>
-          <Button size={'xs'} color={'gray'}>
-            Windows
-          </Button>
-        </Menu.Target>
-        <Menu.Dropdown>
-          <NewWindowMenu />
-        </Menu.Dropdown>
-      </MenuWrapper>
-
-      <MenuWrapper>
-        <Menu.Target>
-          <Button size={'xs'} color={'gray'}>
-            View
-          </Button>
-        </Menu.Target>
-        <Menu.Dropdown>
-          <TaskBarMenuChoices
-            visibleMenuItems={visibleMenuItems}
-            setVisibleMenuItems={setVisibleMenuItems}
-          />
-          <Menu.Item leftSection={<SaveIcon />}>Load/Save Layout</Menu.Item>
-          <Menu.Divider />
-          <Menu.Item leftSection={<VisibilityIcon />}>User Visibility</Menu.Item>
-        </Menu.Dropdown>
-      </MenuWrapper>
+    <Group gap={'xs'} h={30} bg={alpha('var(--mantine-color-dark-5)', 0.9)}>
+      <FileMenu />
+      <WindowsMenu />
+      <ViewMenu
+        visibleMenuItems={visibleMenuItems}
+        setVisibleMenuItems={setVisibleMenuItems}
+      />
       <HelpMenu />
     </Group>
   );
